+++ package/core/rollup.config.mjs
@@ -41,6 +41,7 @@ export default [
       "@react-native-async-storage/async-storage",
       "react-native-mask-input",
       "react-native-svg",
+      "@react-navigation/native",
     ],
   },
   {